refactor: Factorize code, remove unnecessary code, rename parser.py to config_parser.py, rename config_parser.py to parsing_manager, rename tools.py to common_cli_parsing_manager.py
